+++ src/lsp/main.ts
@@ -6,13 +6,16 @@ import { provideClojureDefinition } from '../providers/definition';
 import { setStateValue, getStateValue } from '../../out/cljs-lib/cljs-lib';
 import { downloadClojureLsp } from './download';
 import { readVersionFile, getClojureLspPath } from './utilities';
+import * as os from 'os';
+import * as path from 'path';
 
 const LSP_CLIENT_KEY = 'lspClient';
+const RESOLVE_MACRO_AS_COMMAND = 'resolve-macro-as';
 
 function createClient(clojureLspPath: string): LanguageClient {
     const serverOptions: ServerOptions = {
-        run: { command: clojureLspPath},
-        debug: { command: clojureLspPath},
+        run: { command: clojureLspPath },
+        debug: { command: clojureLspPath },
     };
     const clientOptions: LanguageClientOptions = {
         documentSelector: [{ scheme: 'file', language: 'clojure' }],
@@ -158,7 +161,19 @@ const clojureLspCommands: ClojureLspCommand[] = [
         command: 'server-info',
         category: 'calva.diagnostics'
     }
-]
+];
+
+function sendCommandRequest(command: string, args: (number | string)[]): void {
+    const client = getStateValue(LSP_CLIENT_KEY);
+    if (client) {
+        client.sendRequest('workspace/executeCommand', {
+            command,
+            arguments: args
+        }).catch(e => {
+            console.error(e);
+        });
+    }
+}
 
 function registerLspCommand(client: LanguageClient, command: ClojureLspCommand): vscode.Disposable {
     const category = command.category ? command.category : 'calva.refactor';
@@ -173,23 +188,61 @@ function registerLspCommand(client: LanguageClient, command: ClojureLspCommand):
             const params = [docUri, line, column];
             const extraParam = command.extraParamFn ? await command.extraParamFn() : undefined;
             if (!command.extraParamFn || command.extraParamFn && extraParam) {
-                client.sendRequest('workspace/executeCommand', {
-                    'command': command.command,
-                    'arguments': extraParam ? [...params, extraParam] : params
-                }).catch(e => {
-                    console.error(e);
-                });
+                sendCommandRequest(command.command, (extraParam ? [...params, extraParam] : params));
             }
         }
     });
 }
 
+async function codeLensReferencesHandler(_, line, character): Promise<void> {
+    vscode.window.activeTextEditor.selection = new vscode.Selection(line - 1, character - 1, line - 1, character - 1);
+    await vscode.commands.executeCommand('editor.action.referenceSearch.trigger');
+}
+
+
+async function resolveMacroAsCodeActionCommandHandler(document: string, line: number, character: number): Promise<void> {
+    const macroToResolveAs = await vscode.window.showQuickPick([
+        'clojure.core/def',
+        'clojure.core/defn',
+        'clojure.core/let',
+        'clojure.core/for',
+        'clojure.core/->',
+        'clojure.core/->>',
+        'clj-kondo.lint-as/def-catch-all'
+    ]);
+    const workspaceFolders = vscode.workspace.workspaceFolders;
+    const rootWorkspaceFolder = workspaceFolders && workspaceFolders[0];
+    const homeDirectory = os.homedir();
+    const cljKondoUserConfig = path.join(homeDirectory, '.config', 'clj-kondo', 'config.edn');
+    const configPaths = [cljKondoUserConfig];
+    if (rootWorkspaceFolder) {
+        const cljKondoProjectConfig = path.join(rootWorkspaceFolder.uri.fsPath, '.clj-kondo', 'config.edn');
+        configPaths.push(cljKondoProjectConfig);
+    }
+    const cljKondoConfigPath = await vscode.window.showQuickPick(configPaths, { placeHolder: 'Select where this setting should be saved:' });
+    if (macroToResolveAs && cljKondoConfigPath) {
+        const args = [document, line, character, macroToResolveAs, cljKondoConfigPath];
+        sendCommandRequest(RESOLVE_MACRO_AS_COMMAND, args);
+    }
+}
+
+function resolveMacroAsCommandHandler(): void {
+    const activeTextEditor = vscode.window.activeTextEditor;
+    if (activeTextEditor && activeTextEditor.document && activeTextEditor.document.languageId === 'clojure') {
+        const documentUri = decodeURIComponent(activeTextEditor.document.uri.toString());
+        const { line, character } = activeTextEditor.selection.active;
+        resolveMacroAsCodeActionCommandHandler(documentUri, line + 1, character + 1);
+    }
+}
+
 function registerCommands(context: vscode.ExtensionContext, client: LanguageClient) {
-    // The title of this command is dictated by clojure-lsp and is executed when the user clicks the references code lens above a symbol
-    context.subscriptions.push(vscode.commands.registerCommand('code-lens-references', async (_, line, character) => {
-        vscode.window.activeTextEditor.selection = new vscode.Selection(line - 1, character - 1, line - 1, character - 1);
-        await vscode.commands.executeCommand('editor.action.referenceSearch.trigger');
-    }));
+    // The title of this command is dictated by clojure-lsp and is executed when the user clicks the references code lens for a symbol
+    context.subscriptions.push(vscode.commands.registerCommand('code-lens-references', codeLensReferencesHandler));
+
+    // The title of this command is dictated by clojure-lsp and is executed when the user executes the Resolve Macro As code action
+    context.subscriptions.push(vscode.commands.registerCommand(RESOLVE_MACRO_AS_COMMAND, resolveMacroAsCodeActionCommandHandler));
+
+    context.subscriptions.push(vscode.commands.registerCommand('calva.linting.resolveMacroAs', resolveMacroAsCommandHandler));
 
     context.subscriptions.push(
         ...clojureLspCommands.map(command => registerLspCommand(client, command))